remove unused imports
add logging

fix async logic
