+++ packages/server/lambdas/alfresco/fileRequestBuilder/alfrescoRequestBuilder.ts
@@ -1,16 +1,11 @@
 import axios, { AxiosError, AxiosResponse } from 'axios';
 import { getAlfrescoUrlBase } from '../../../utils/alfresco';
-import { FileDeleteRequest, FileDeleteRequestBody, IFileRequestBody, IFileResponse } from './types';
+import { FileDeleteRequest, FileDeleteRequestBody, IFileResponse } from './types';
 import { FormData } from 'formdata-node';
-import fetch from 'node-fetch';
+import { log } from '../../../utils/logger';
 
-const dataToBlob = async (data: string) => {
-  const base64 = await fetch(data);
-  const blob = base64.blob();
-  return blob;
-};
 const postFile = async (data: FormData, nodeId: string) => {
-  await axios({
+  return await axios({
     method: 'post',
     url: `${getAlfrescoUrlBase()}/nodes/${nodeId}/children`,
     data: data,
@@ -21,27 +16,21 @@ const postFile = async (data: FormData, nodeId: string) => {
 export class AlfrescoFileRequestBuilder {
   public async requestBuilder(
     nodeId: string,
-    requestBody: string,
-  ): Promise<AxiosResponse<IFileResponse> | AxiosError | void> {
-    dataToBlob(requestBody)
-      .then((blob) => {
-        const bodyFormData = new FormData();
+    requestBody: FormData,
+  ): Promise<AxiosResponse<IFileResponse> | AxiosError> {
+    log.info('requestBuilder 23');
 
-        bodyFormData.append('filedata', blob, Date.now().toString() + '_FILE.txt');
-
-        postFile(bodyFormData, nodeId)
-          .then((response) => {
-            console.log('RESPONSE1: ', response);
-            return response;
-          })
-          .catch((error: AxiosError) => {
-            console.log(error);
-            return error;
-          });
+    const response = postFile(requestBody, nodeId)
+      .then((response) => {
+        console.log('RESPONSE1: ', response);
+        return response;
       })
-      .catch((error) => {
+      .catch((error: AxiosError) => {
+        console.log('ERROR1', error);
         return error;
       });
+
+    return response;
   }
   public deleteRequestBuilder(requestParameters: FileDeleteRequestBody): FileDeleteRequest {
     throw new Error('Method not implemented.');
